requirements.txt
streamlit>=1.31.0
Pillow>=9.5.0
google-generativeai>=0.3.1
backoff>=2.2.1